Add list rental class.
